refactor to a datamodel?
